refactor: handle initial node and initial properties in mixin
