+++ packages/core/index.js
@@ -21,8 +21,6 @@ export { bootstrap } from './src/lib/bootstrap.js';
 export { define } from './src/lib/define.js';
 export { render } from './src/lib/render.js';
 
-const Component = proxy(class {});
-
 /**
  * Simple Custom Component with some behaviors.
  * @class BaseComponent
@@ -53,25 +51,10 @@ const Component = proxy(class {});
  * }
  * ```
  */
-export class BaseComponent extends mix(Component).with(
+export class BaseComponent extends mix(proxy(class {})).with(
     MIXINS.ComponentMixin,
     MIXINS.PropertiesMixin,
     MIXINS.StyleMixin,
     MIXINS.EventsMixin,
     MIXINS.TemplateMixin
-) {
-    constructor(node) {
-        super();
-        if (!node) {
-            let desc = registry.get(this.is, true);
-            let config = desc.config;
-            node = document.createElement(
-                config.extends ? config.extends : desc.is
-            );
-            if (config.extends) {
-                node.setAttribute('is', desc.is);
-            }
-        }
-        this.node = node;
-    }
-}
+) {}
