+++ packages/core/test/units/base.js
@@ -14,7 +14,7 @@ define('test-base-component-2', TestBaseComponent2, {
 });
 
 describe('Unit: BaseComponent', () => {
-    const elem = DOM.createElement('test-base-component');
+    const elem = new TestBaseComponent();
     elem.lastName = 'Turing';
 
     describe('> created', () => {
@@ -56,7 +56,7 @@ describe('Unit: BaseComponent', () => {
 });
 
 describe('Unit: BaseComponent with native element', () => {
-    const elem = DOM.createElement('test-base-component-2');
+    const elem = new TestBaseComponent2();
     elem.lastName = 'Turing';
 
     describe('> created', () => {
