+++ packages/custom-elements-v0/src/lib/render.js
@@ -1,8 +1,5 @@
 export function render(node, Component, props = {}) {
-    let element = new Component();
-    for (let k in props) {
-        element[k] = props[k];
-    }
+    let element = new Component(props);
     node.appendChild(element);
     return element;
 }
